financial-advisor/pyproject.toml
[project]
name = "financial-advisor"
version = "0.1"
description = "AI-driven agent designed to facilitate the exploration of the financial advisor landscape"
authors = [{ name = "Antonio Gulli", email = "gulli@google.com" }]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
google-cloud-aiplatform = { version = "^1.93.0", extras = [
    "adk",
    "agent-engines",
] }
python = ">=3.10, <3.13"
python-dotenv = "^1.0.1"
google-adk = { version = ">=1.15.1", extras = ["a2a"] }
torch = "^2.5.0"
stable-baselines3 = "^2.4.0"
opentelemetry-api = "1.37.0"
opentelemetry-sdk = "1.37.0"
opentelemetry-exporter-otlp = "1.37.0"
google-cloud-iam = "^2.20.0"
google-cloud-secret-manager = "^2.25.0"
google-auth = "^2.43.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
pytest-asyncio = ">=0.23.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"   ]
build-backend = "poetry.core.masonry.api"
